requirements.txt
fastapi==0.115.0
openai==1.47.0
python-dotenv==1.0.1
uvicorn==0.30.6